@@ -57,7 +57,7 @@ class MyUpgradeAlertState extends UpgradeAlertState {
         textButtonTheme: const TextButtonThemeData(
           style: ButtonStyle(
             // Change the color of the text buttons.
-            foregroundColor: MaterialStatePropertyAll(Colors.orange),
+            foregroundColor: WidgetStatePropertyAll(Colors.orange),
           ),
         ),
       ),
