@@ -7,7 +7,7 @@ import Foundation
 
 import device_info_plus
 import package_info_plus
-import path_provider_macos
+import path_provider_foundation
 import shared_preferences_foundation
 import url_launcher_macos
 
