@@ -8,7 +8,6 @@ import 'package:version/version.dart';
 import 'package:http/http.dart' as http;
 
 class ITunesSearchAPI {
-  /// iTunes Search API documentation URL
   final String iTunesDocumentationURL =
       'https://affiliate.itunes.apple.com/resources/documentation/itunes-store-web-service-search-api/';
 
@@ -32,14 +31,18 @@ class ITunesSearchAPI {
   /// ```lookupURLByBundleId('com.google.Maps');```
   /// ```lookupURLByBundleId('com.google.Maps', country: 'FR');```
   Future<Map?> lookupByBundleId(String bundleId,
-      {String? country = 'US', bool useCacheBuster = true}) async {
+      {String? country = 'US',
+      String? language = 'en',
+      bool useCacheBuster = true}) async {
     assert(bundleId.isNotEmpty);
     if (bundleId.isEmpty) {
       return null;
     }
 
     final url = lookupURLByBundleId(bundleId,
-        country: country ?? '', useCacheBuster: useCacheBuster)!;
+        country: country ?? '',
+        language: language ?? '',
+        useCacheBuster: useCacheBuster)!;
     if (debugLogging) {
       print('upgrader: download: $url');
     }
@@ -94,14 +97,18 @@ class ITunesSearchAPI {
   /// ```lookupURLByBundleId('com.google.Maps');```
   /// ```lookupURLByBundleId('com.google.Maps', country: 'FR');```
   String? lookupURLByBundleId(String bundleId,
-      {String country = 'US', bool useCacheBuster = true}) {
+      {String country = 'US',
+      String language = 'en',
+      bool useCacheBuster = true}) {
     if (bundleId.isEmpty) {
       return null;
     }
 
-    return lookupURLByQSP(
-        {'bundleId': bundleId, 'country': country.toUpperCase()},
-        useCacheBuster: useCacheBuster);
+    return lookupURLByQSP({
+      'bundleId': bundleId,
+      'country': country.toUpperCase(),
+      'lang': language
+    }, useCacheBuster: useCacheBuster);
   }
 
   /// Look up URL by id.
