@@ -40,8 +40,10 @@ class UpgraderAppStore extends UpgraderStore {
     iTunes.debugLogging = state.debugLogging;
     iTunes.client = state.client;
     iTunes.clientHeaders = state.clientHeaders;
-    final response = await (iTunes
-        .lookupByBundleId(state.packageInfo!.packageName, country: country));
+    final response = await (iTunes.lookupByBundleId(
+        state.packageInfo!.packageName,
+        country: country,
+        language: language));
 
     if (response != null) {
       final version = iTunes.version(response);
